# This is a combination of 2 commits. # The first commit's message is:
Implement and use async_io and larency measuring

async_io - is the interface for asynchronous input/output
to sockets based on libev library.
latency - is a gap between request to a server and
receiving an answer from it.
Latency can be measured in seconds, microseconds and
milliseconds.

# This is the 2nd commit message:

Fix latency measuring, add ability to choosing RPS